initial commit. rename package due to #3
added initial gradle structure and renamed package.
