+++ build.gradle
@@ -1,0 +1,3 @@
+apply plugin: 'idea'
+
+apply plugin: 'java'
